Task2/parser_module_AG/requirements.txt
pdfminer.six
python-docx
reportlab
